fix(ci): split static files test to pass ci
CI hangs on python 3.7 (don't know why). Some research was done without satisfying result (see conversation in #134 and #137). Splitting test into two separate files fixed this issue.
